fix: Allow docker build to use the go-cache
Go's build correctness is world class. Let the Go toolchain leverage the
go build cache that's being saved as part of the max mode buildx layer
caching. Otherwise, you're eating the cost of pulling and pushing all of
that for nothing and that's about 399M which is quite a bit.